refactor: 将编辑前须知移出为一个文件
